# src/main/java/dev/alkolhar/servdesk/ticket/TicketModel.java
package dev.alkolhar.servdesk.ticket;

import java.time.Instant;
import org.jspecify.annotations.Nullable;
import org.springframework.hateoas.RepresentationModel;

@SuppressWarnings("NotNullFieldNotInitialized")
public class TicketModel extends RepresentationModel<TicketModel> {

	private Long id;
	private String ticketNumber;
	private TicketType type;
	private TicketStatus status;
	private String subject;
	private @Nullable String description;
	private @Nullable Long categoryId;
	private @Nullable Long priorityId;
	private Long requesterId;
	private @Nullable Long assigneeId;
	private @Nullable Long teamId;
	private @Nullable Instant resolvedAt;
	private @Nullable Instant closedAt;
	private Instant createdAt;
	private Instant updatedAt;
	private @Nullable String createdBy;
	private @Nullable String updatedBy;

	public Long getId() {
		return id;
	}

	public void setId(Long id) {
		this.id = id;
	}

	public String getTicketNumber() {
		return ticketNumber;
	}

	public void setTicketNumber(String ticketNumber) {
		this.ticketNumber = ticketNumber;
	}

	public TicketType getType() {
		return type;
	}

	public void setType(TicketType type) {
		this.type = type;
	}

	public TicketStatus getStatus() {
		return status;
	}

	public void setStatus(TicketStatus status) {
		this.status = status;
	}

	public String getSubject() {
		return subject;
	}

	public void setSubject(String subject) {
		this.subject = subject;
	}

	public @Nullable String getDescription() {
		return description;
	}

	public void setDescription(@Nullable String description) {
		this.description = description;
	}

	public @Nullable Long getCategoryId() {
		return categoryId;
	}

	public void setCategoryId(@Nullable Long categoryId) {
		this.categoryId = categoryId;
	}

	public @Nullable Long getPriorityId() {
		return priorityId;
	}

	public void setPriorityId(@Nullable Long priorityId) {
		this.priorityId = priorityId;
	}

	public Long getRequesterId() {
		return requesterId;
	}

	public void setRequesterId(Long requesterId) {
		this.requesterId = requesterId;
	}

	public @Nullable Long getAssigneeId() {
		return assigneeId;
	}

	public void setAssigneeId(@Nullable Long assigneeId) {
		this.assigneeId = assigneeId;
	}

	public @Nullable Long getTeamId() {
		return teamId;
	}

	public void setTeamId(@Nullable Long teamId) {
		this.teamId = teamId;
	}

	public @Nullable Instant getResolvedAt() {
		return resolvedAt;
	}

	public void setResolvedAt(@Nullable Instant resolvedAt) {
		this.resolvedAt = resolvedAt;
	}

	public @Nullable Instant getClosedAt() {
		return closedAt;
	}

	public void setClosedAt(@Nullable Instant closedAt) {
		this.closedAt = closedAt;
	}

	public Instant getCreatedAt() {
		return createdAt;
	}

	public void setCreatedAt(Instant createdAt) {
		this.createdAt = createdAt;
	}

	public Instant getUpdatedAt() {
		return updatedAt;
	}

	public void setUpdatedAt(Instant updatedAt) {
		this.updatedAt = updatedAt;
	}

	public @Nullable String getCreatedBy() {
		return createdBy;
	}

	public void setCreatedBy(@Nullable String createdBy) {
		this.createdBy = createdBy;
	}

	public @Nullable String getUpdatedBy() {
		return updatedBy;
	}

	public void setUpdatedBy(@Nullable String updatedBy) {
		this.updatedBy = updatedBy;
	}
}